emails: account mail + Mailer back to core; billing/growth mail stays ee (#475)
Fountain.Mailer and the account templates (verification, password reset,
suspension x2, deletion, email-change x2) return to apps/fountain — mail a
community instance needs must never sit behind a future ee license. The
seven billing/growth templates (welcome, trial x2, payment x4) move to a
new Fountain.Emails.BillingEmails in ee/, which borrows the now-public
UserEmails.from_address/0 and support_phrase/0 so the whole mail surface
keeps one sender and one support-contact policy.

Callers updated: welcome/trial-ending/lifecycle workers (+ their tests)
alias BillingEmails; account/email-change/verification workers and the
password-reset controller keep UserEmails untouched. Test file split the
same way; coveralls skip path for mailer.ex restored to core.

Closes #475.
