+++ apps/fountain/test/fountain/emails/user_emails_test.exs
@@ -40,36 +40,6 @@ defmodule Fountain.Emails.UserEmailsTest do
     end
   end
 
-  describe "deliver_welcome_email/1" do
-    test "welcomes the user with an onboarding link" do
-      user = insert_verified_user()
-
-      assert {:ok, _email} = UserEmails.deliver_welcome_email(user)
-
-      assert_email_sent(fn email ->
-        assert email.subject == "Welcome to Fountain"
-        assert email.to == [{user.email, user.email}]
-        assert email.html_body =~ "/onboarding/step_1"
-        assert email.text_body =~ "/onboarding/step_1"
-      end)
-    end
-
-    test "links to the dashboard instead once onboarding is complete" do
-      user = insert_verified_user()
-      now = DateTime.utc_now() |> DateTime.truncate(:second)
-
-      {:ok, user} =
-        user |> Ecto.Changeset.change(onboarding_completed_at: now) |> Repo.update()
-
-      assert {:ok, _email} = UserEmails.deliver_welcome_email(user)
-
-      assert_email_sent(fn email ->
-        refute email.text_body =~ "/onboarding"
-        assert email.subject == "Welcome to Fountain"
-      end)
-    end
-  end
-
   describe "link absoluteness" do
     # These assert on the scheme, not just the path. The original tests only
     # checked "/users/confirm/<token>", which passed happily while production
