+++ ee/test/fountain/emails/billing_emails_test.exs
@@ -1,0 +1,37 @@
+defmodule Fountain.Emails.BillingEmailsTest do
+  use Fountain.DataCase, async: true
+
+  import Swoosh.TestAssertions
+
+  alias Fountain.Emails.BillingEmails
+
+  describe "deliver_welcome_email/1" do
+    test "welcomes the user with an onboarding link" do
+      user = insert_verified_user()
+
+      assert {:ok, _email} = BillingEmails.deliver_welcome_email(user)
+
+      assert_email_sent(fn email ->
+        assert email.subject == "Welcome to Fountain"
+        assert email.to == [{user.email, user.email}]
+        assert email.html_body =~ "/onboarding/step_1"
+        assert email.text_body =~ "/onboarding/step_1"
+      end)
+    end
+
+    test "links to the dashboard instead once onboarding is complete" do
+      user = insert_verified_user()
+      now = DateTime.utc_now() |> DateTime.truncate(:second)
+
+      {:ok, user} =
+        user |> Ecto.Changeset.change(onboarding_completed_at: now) |> Repo.update()
+
+      assert {:ok, _email} = BillingEmails.deliver_welcome_email(user)
+
+      assert_email_sent(fn email ->
+        refute email.text_body =~ "/onboarding"
+        assert email.subject == "Welcome to Fountain"
+      end)
+    end
+  end
+end
